fix: three unbounded reactions to a number a peer chose, found by fuzzing them
The three fuzz targets in the tree predated the coded path, so nothing had
ever fed arbitrary bytes to the parsers the erasure work added. Seven targets
now cover the decoders that read what a peer sends: the coded datagram header
and the assembler under it, the sliding-window decoder, ACK range lists, the
UDP packet frame and its destination, and both SOCKS5 parsers. They found
four defects, three of them the same shape.

**A window walked to wherever it was told.** fec's decoder advances its oldest
identifier by evicting one at a time, because each eviction is a symbol the
layer above has to be told will never arrive. The identifier comes from four
bytes of a datagram: one source symbol naming an ESI 2^30 ahead walked the
whole distance, scanning the row set and appending a Lost entry per step, with
the receiving path's lock held. It never returned. Only a window's worth can
ever have been admitted, so the walk is now bounded to a window and the rest
of the jump is one step.

**A loss estimator counted out four billion losses.** Same shape, one layer
over: the estimator retires sequence numbers one at a time until the undecided
range fits its ring, and the sequence number is also the peer's. Two datagrams,
one numbered near zero and one near 2^32, and the receive loop is gone. A gap
larger than any real burst -- 8192 consecutive losses is a path that has
stopped, not one that is dropping -- is now a discontinuity in the numbering
rather than a measurement of it: the estimator restarts from where it is, and
counts that it did, so a path losing packets can still be told from a peer that
is not numbering them.

Both are reachable from one authenticated peer, on the receive loop, under the
lock, with no recover anywhere above them.

**A frame size that is negative on a 32-bit build.** parseFrames converted a
uint32 straight to int and compared it against the remaining length. On 32-bit
that wraps negative for anything past 2 GiB, the comparison passes, and the
slice panics. The bound was only accidentally safe on 64-bit; the seed corpus
catches it under GOARCH=386 and passes after.

**A canonicalizer that returned what it was given.** DecodeDestination
computed the canonical form, used it to decide the input was valid, then
returned the input. So "host:0000443" was accepted as itself and one
destination had as many spellings as a peer cared to write. Fixing that
exposed the encoder was not idempotent either: "[ ]:1" canonicalized to " :1",
which it then rejected, because the TrimSpace at the front ate a host that
SplitHostPort had legitimately unbracketed. A host is never whitespace or
control characters, and saying so makes the function a fixed point of itself.

Nothing keys on the destination string today and the dial policy re-parses and
checks resolved addresses rather than matching text, so this was not a way
past it.

+++ internal/coded/coded.go
@@ -637,9 +637,16 @@ func (a *assembler) prune() {
 
 func parseFrames(payload []byte, out [][]byte) [][]byte {
 	for len(payload) >= frameHeader {
-		size := int(binary.BigEndian.Uint32(payload))
+		size := binary.BigEndian.Uint32(payload)
 		payload = payload[frameHeader:]
-		if size > len(payload) {
+		// Compared as a uint64 rather than converted to int first. On a
+		// 32-bit build int(uint32) wraps negative for anything past 2 GiB,
+		// a negative size passes a `size > len(payload)` test, and the slice
+		// below panics on a negative bound -- so four bytes off the wire
+		// take down a receive loop that has no recover above it. The bound
+		// is only accidentally safe on 64-bit, which is not a property to
+		// leave a wire parser resting on.
+		if uint64(size) > uint64(len(payload)) {
 			break
 		}
 		out = append(out, append([]byte(nil), payload[:size]...))
